requirements.txt
solana>=0.34.0
solders>=0.21.0
httpx>=0.27.0
websockets>=12.0
python-dotenv
base58
certifi
rich
